backend/requirements-seed.txt
# Requirements for PRM Service Database Seeder
# Install with: pip install -r requirements-seed.txt

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
loguru>=0.7.0
python-dotenv>=1.0.0
